BAN-109 add new components in ProfileSettings
