+++ src/widgets/profileSettings/components/accountManagement/AccountManagement.tsx
@@ -1,0 +1,3 @@
+export const AccountManagement = () => {
+  return <div>AccountManagement</div>
+}
