+++ src/widgets/profileSettings/components/devices/Devices.tsx
@@ -1,0 +1,3 @@
+export const Devices = () => {
+  return <div>Devices</div>
+}
