+++ src/widgets/profileSettings/components/index.ts
@@ -1,0 +1,4 @@
+export * from './generalInformation/GeneralInformation'
+export * from './devices/Devices'
+export * from './accountManagement/AccountManagement'
+export * from './myPayments/MyPayments'
