+++ src/widgets/profileSettings/components/myPayments/MyPayments.tsx
@@ -1,0 +1,3 @@
+export const MyPayments = () => {
+  return <div>MyPayments</div>
+}
