+++ src/widgets/profileSettings/model/hooks/index.ts
@@ -1,1 +1,1 @@
-export * from './useProfileSettings'
+export * from './useGeneralInformation'
