+++ src/widgets/profileSettings/model/schemas/index.ts
@@ -1,1 +1,1 @@
-export * from './profileSettingsSchema'
+export * from './generalInformationSchema'
