+++ src/widgets/profileSettings/ui/ProfileSettings.module.scss
@@ -4,79 +4,3 @@
   padding-left: 25px;
   margin: 0 auto;
 }
-
-.wrapper {
-  margin-top: 25px;
-  padding-bottom: 25px;
-  display: flex;
-  gap: 40px;
-  border-bottom: 1px solid #333;
-}
-
-.image {
-  max-width: 202px;
-}
-
-.img {
-  width: 192px;
-  height: 192px;
-  background-color: #171717;
-  border-radius: 50%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-bottom: 25px;
-  &__btn {
-    white-space: nowrap;
-  }
-  &__icon {
-    width: 48px;
-    height: 48px;
-  }
-}
-
-.info {
-  width: 100%;
-}
-
-.username,
-.firstname,
-.lastname,
-.date {
-  margin-bottom: 25px;
-}
-
-.text {
-  color: #8d9094;
-  font-size: 15px;
-  margin-top: 25px;
-}
-
-.selectors {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 20px;
-  margin-bottom: 25px;
-}
-
-.selector {
-  width: 100%;
-}
-
-.select {
-  width: 100%;
-}
-
-.btn {
-  margin-top: 25px;
-  text-align: right;
-}
-
-@media (max-width: 1000px) {
-  .wrapper {
-    display: block;
-  }
-  .image {
-    margin: 0 auto;
-  }
-}
