+++ src/widgets/profileSettings/ui/ProfileSettings.tsx
@@ -1,12 +1,14 @@
 'use client'
 
-import React from 'react'
+import React, { useState } from 'react'
 import { Tabs } from '@/shared/ui/tabs/Tabs'
 import styles from './ProfileSettings.module.scss'
-import { ImageOutline } from '@/assets/icons/components'
-import { Button, DatePicker, Input, Textarea } from '@/shared/ui'
-import Select from '@/shared/ui/select/Select'
-import { useProfileSettings } from '@widgets/profileSettings/model/hooks'
+import {
+  AccountManagement,
+  Devices,
+  GeneralInformation,
+  MyPayments,
+} from '@widgets/profileSettings/components'
 
 export default function ProfileSettings() {
   const tabs = [
@@ -16,106 +18,34 @@ export default function ProfileSettings() {
     { value: 'tab-4', title: 'My payments' },
   ]
 
-  const countryOptions = [
-    { label: 'Kiribati', value: 'kiribati' },
-    { label: 'Bhutan', value: 'bhutan' },
-    { label: 'Djibouti', value: 'djibouti' },
-  ]
+  const [activeTab, setActiveTab] = useState('tab-1')
 
-  const cityOptions = [
-    { label: 'Asunción', value: 'asuncion' },
-    { label: 'Guadalajara', value: 'guadalajara' },
-    { label: 'Kalkuta', value: 'kalkuta' },
-  ]
-
-  const {
-    username,
-    setUsername,
-    firstName,
-    setFirstName,
-    lastName,
-    setLastName,
-    country,
-    setCountry,
-    city,
-    setCity,
-    aboutMe,
-    setAboutMe,
-    isFormValid,
-    handleSubmit,
-  } = useProfileSettings()
+  const renderTabContent = () => {
+    switch (activeTab) {
+      case 'tab-1':
+        return <GeneralInformation />
+      case 'tab-2':
+        return <Devices />
+      case 'tab-3':
+        return <AccountManagement />
+      case 'tab-4':
+        return <MyPayments />
+      default:
+        return null
+    }
+  }
 
   return (
     <div className={styles.container}>
       <div className={styles.header}>
-        <Tabs className={styles.tabs} tabs={tabs} defaultValue="tab-1" />
+        <Tabs
+          className={styles.tabs}
+          tabs={tabs}
+          defaultValue="tab-1"
+          onValueChange={setActiveTab}
+        />
       </div>
-      <div className={styles.inner}>
-        <div className={styles.wrapper}>
-          <div className={styles.image}>
-            <div className={styles.img}>
-              <ImageOutline className={styles.img__icon} />
-            </div>
-            <Button className={styles.img__btn} variant="outline">
-              Select Profile Photo
-            </Button>
-          </div>
-          <div className={styles.info}>
-            <Input
-              className={styles.username}
-              subTitle="Username *"
-              value={username}
-              onChange={e => setUsername(e.target.value)}
-            />
-            <Input
-              className={styles.firstname}
-              subTitle="First Name *"
-              value={firstName}
-              onChange={e => setFirstName(e.target.value)}
-            />
-            <Input
-              className={styles.lastname}
-              subTitle="Last Name *"
-              value={lastName}
-              onChange={e => setLastName(e.target.value)}
-            />
-            <DatePicker id="date-picker-with-required-field" label="Date of birth" />
-            <div className={styles.selectors}>
-              <div className={styles.selector}>
-                <p className={styles.text}>Select your country</p>
-                <Select
-                  className={styles.select}
-                  options={countryOptions}
-                  value={country}
-                  onValueChange={setCountry}
-                  placeholder="Select country"
-                />
-              </div>
-              <div className={styles.selector}>
-                <p className={styles.text}>Select your city</p>
-                <Select
-                  className={styles.select}
-                  options={cityOptions}
-                  value={city}
-                  onValueChange={setCity}
-                  placeholder="Select city"
-                />
-              </div>
-            </div>
-            <Textarea
-              className={styles.textarea}
-              title="About Me"
-              value={aboutMe}
-              onChange={e => setAboutMe(e.target.value)}
-            />
-          </div>
-        </div>
-        <div className={styles.btn}>
-          <Button variant="primary" disabled={!isFormValid} onClick={handleSubmit}>
-            Save Changes
-          </Button>
-        </div>
-      </div>
+      {renderTabContent()}
     </div>
   )
 }
